crash: Add function to kernel warning basenames.
If two kernel warnings are generated in quick succession, the second
won't necessarily be written (filenames might collide), so the
second warning won't be collected.

This could cause us to always miss the second warning if two warnings are always
logged in quick succession for a particular problem.

It also causes the platform.KernelWarning tast test to flake if a kernel
warning comes in right before the artificial one that it generates.

BUG=chromium:1066939
TEST=Manually induce kernel warning, check generated filenames and meta\
     file contents.
TEST=tast run $DUT platform.KernelWarning.*

Cq-Depend: chromium:2132749
Change-Id: Ifee7b601fc9ec597cd0c10bd5bcbad722bdcd484
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2132829
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>