adbd: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes it clearer
that the above code is an infinite loop. If someone does add a `break`
down the line, they'll get a -Werror'ed warning about lacking a
`return;` with a valid value.

BUG=chromium:1066998
TEST=CQ

Change-Id: I64b59867be231dabc5eae4606539b0afd0a16129
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2132920
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>