camera: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes this code
marginally clearer, since the above code is an infloop (only exited with
`return`s). `int main()` always gets an implicit `return 0;` at the
bottom, so this should be harmless.

BUG=chromium:1066998
TEST=CQ

Change-Id: I97b8e7cb89676a55a5269bf9c6dfbb08082ca0f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2132921
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>