shill: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes this code
clearer and easier to modify, since if a case is added to this `switch`
later, the user should get a -Werror'ed warning if they forget to add a
`return`.

BUG=chromium:1066998
TEST=CQ

Change-Id: I10b2931f4d70cab494c4d936786b7db75474fae8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2133014
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>