vm_tools: concierge: allow xattr to exist when resizing
Change the fsetxattr() call to not return an error if the xattr already
exists.  With XATTR_CREATE, the call would fail if the xattr was already
in place from a previous resize.  With flags == 0, the call will create
or update the xattr as needed.

BUG=chromium:1067023
TEST=`vmc resize termina <size>` twice with two different sizes

Change-Id: I4e69127d1dacc600d6f6521d432dde69884f21fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2133209
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>