arc: networkd: Pre-create TAP devices for ARCVM multinet.
This patch enables multiple device support for ARCVM.
The impl now accepts a list of device configurations when it is built,
that will be empty for single networking or populated for multi-net.
Now, when the impl starts up, after creating the arc device
as usual, for each configuration, a TAP device is opened and associated.
A prior patch enabled dynamic binding of configurations to devices as
they are added/removed. Together now, devices can be added and removed
and "connected/disconnected" to/from existing TAPs.

This patch does not change any current behavior - multinet is still
disabled.

BUG=b:136417971
TEST=units, tast arc.Boot.vm, arc.Downloads.vm

Change-Id: I71987d4e35b6e148ef4c67ca6b08f1b3ebc7bd39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2134001
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>