lorgnette: Inherit supplemental groups
lorgnette needs to be able to access USB devices to operate scanners.
In order to avoid running as root, we have a list of known scanners that
udev changes to the scanner group.  Some  scanners are also IPP over USB
devices, which then get changed to the ippusb group instead.  Since
saned is in the ippusb group, keep supplemental groups in lorgnette's
minijail so that it can access both types of devices.

BUG=b:153108022
TEST=TR8500 shows up in scanner listing; page can be scanned.

Cq-Depend: chromium:2135854
Change-Id: I2ae471a4133fec2388aed46e7c945ba187f50d93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2135857
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>