diagnostics: add ProbeError to CpuInfo telemetry probe
This change modifies the CpuInfo probe to be a mojo union between the
CpuInfo and a ProbeError. This insures the result of the probe can only
be the data or an error. In addition to the data structure modification,
add ProbeErrors for invalid conditions that were previously ignored.

BUG=chromium:1041153
TEST=FEATURES=test emerge-sarien diagnostics (pass);
(dut) telem --category=cpu (shows CPU info)

Change-Id: I23e55b1c95fdcb5170eb9706a2e0f33abcc4cb7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2135858
Commit-Queue: Trent Begin <tbegin@chromium.org>
Tested-by: Trent Begin <tbegin@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>