dlcservice: Shuffle the DlcService tests around
We need to have a base test class to setup the DLC service
environment. Currently, the only class we have is the DlcServiceTest
class. But if we want to test other classes, then we have to rewrite
most of the things again. This CL prepares the DlcServicseTest into
BaseTest so it can be moved to some test util file and be used by other
tests.

Also marks the Third-Dlc as preload-allowed so we don't have to have a
separate imageloader.json for it.

BUG=chromium:1042987
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: Id373ac6d21a9e1097ea256104e589063c97accad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2136071
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>