chaps: replace |is_free_required_| with adaptive |Free|
This CL implements a adaptive |Free| and gets rid of
|is_free_required_|.
The old flag is hard to maintain correctly and the definition
is not clear. By making |Free| able to handle the paring no
matter if it's successful or failed, as well as the case of injected
attributes, we can just correctly release the memory in all cases.

BUG=chromium:1016094
TEST=unittests
TEST=the testcase in the bug link is/isn't reproduced before/after
cros_workon start chaps

Change-Id: Ia831bdaca8c726762040afd8d04182a087f59b68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2138943
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>