hardware_verifier: use mktemp to prevent race condition
In our test, we will periodically check the existence of the result file
and copy it to host.  However, current implementation in the init script
will cause a race condition that only an empty file is copied before
hardware_verifier dumps the report.  This commit fixes the issue by
dumping the report to a temp file and move to the dest filepath.

BUG=b:153293193
TEST=manual test

Change-Id: Ibe4d32b1295bb248e8a995b8075d6701bd092854
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2136992
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>