cros-disks: Fix RAR archives in Guest mode
The source of a RAR archive mount operation might be inside the user's
cryptohome. In Guest mode cros-disks might not have access to that file
if the user session is in a different mount namespace, and the file
exists in a folder that does not exist in cros-disks's mount namespace.

Fix by having RarManager check for the existence of the file in
Chrome's mount namespace, and if that's the case then execute the FUSE
helper in that namespace.

BUG=chromium:1067801
TEST=Unit tests still pass.
TEST=RAR archive mounting in Guest mode works.
TEST=Multipart RAR archive mounting in Guest mode works.
TEST=RAR archive mounting in regular mode works.
TEST=Multipart RAR archive mounting in regular mode works.

Change-Id: I1a7ae6b961d8e3fb6c9f6adcd57ecf634e846fe0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2137503
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>