common-mk: Manually parse verbose argument.
First step to removing the verbose argument in favor of the
new verbose argument in the base parser. This is needed to
uprev some CROS_WORKON_BLACKLISTed packages that depend
on common-mk so the verbose argument can be added in the
base parser after.

BUG=chromium:1071606
TEST=cq

Change-Id: I36fb7fa2ab02f4b737d4acd20f0a0c5b4c3d259b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2137587
Tested-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>