chromeos-config: add a cros_config_mock script
The factory process requires getting values for a "mocked" SKU id. I
was originally planning to add this as a flag to cros_configfs, but
given that tool will be getting a total rewrite here soon, I don't
want to go about adding this functionality to libcros_config only to
go deleting it pretty soon, or worse yet end up having to keep the
code just for this functionality.

So a shell-script seems appropriate here. Hopefully this can be added
to the new cros_configfs command after we test it and it's stable.

To be installed under /usr/bin/cros_config_mock

BUG=b:151296815,b:152349251,b:152291015
TEST=cros_config_mock able to get values for other identities

Change-Id: I00df88d6cbfd62c0b9369eb4c8e5dc68297edbfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2136897
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>