dlcservice: Add dbus_adaptor_test.cc
We were not testing the classes for dbus functions. This CL adds those
tests.

This CL also makes an interface of the DlcService so it can be easily
mocked.

Also changes the DlcGet method of DlcBase to return a const pointer
instead of a const reference. Without this unittesting anything involves
this class would be impossible because the copy constructor has been
deleted.

BUG=chromium:1042987
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: Ie1970163ce2cde2041dd6f99db4456d092ae71c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2138612
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>