system_api: Add constants for chaps
This CL adds a file for saving constants that will be used in chaps and
chrome.

It adds constants that are used to define attributes stored with keys.
Two attributes are added for the moment. CKA_CHROMEOS_KEY_PERMISSIONS is
added to save the contexts in which a key can be used.

CKA_CHROMEOS_BUILTIN_PROVISIONING_PROFILE_ID is added to store the
profile id of a key created by the built-in certificate provisioning
feature.

BUG=chromium:1045895
TEST=Manual testing by using the constants in a testing file.
Change-Id: I2dcfa000fb8ea30dddf55a97c4a0d70fa7d7a887
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2139976
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Omar Morsi <omorsi@google.com>
Commit-Queue: Omar Morsi <omorsi@google.com>