kerberos: refactor config parser enctypes retrieval
Refactor GetEncryptionTypes() to return a boolean error code. This way,
this method will not fail silently and we won't rely on the caller to
send a valid config. An invalid config is not expected to be sent to
GetEncryptionTypes() right now, but it's safer to cover this case.

BUG=chromium:1067874

TEST=cros_workon_make --board=amd64-generic --test kerberos

Change-Id: I8f630b7096101fbbd9097f8e7fc690bb71cbc96d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2139977
Tested-by: Felipe Andrade <fsandrade@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>