hermes: Split D-Bus interface between Manager and Profile
Prior to this change, the D-Bus API consisted of a single interface
that had no properties. While this works to some degree with the
assumption that there will only ever be one client, allowing for
Hermes to be interacted with both through a CLI and the UI
necessitates a more flexible API. In addition, having a Profile
interface with properties allows for clients to interact with Hermes
without having to know about google-lpa protobufs.

BUG=chromium:1058054, chromium:847619
TEST=Manual testing using `dbus-send --system --print-reply \
                          --dest=org.chromium.Hermes $OBJECT \
                          org.freedesktop.DBus.Properties.GetAll \
                          string:"$INTERFACE"`:
      * By default Manager's Profiles property is empty (board has no
      qrtr services so CardQrtr hasn't reached a state where it can
      actually communicate with the eUICC).
      * Adding a fake Profile in the Manager constructor causes the
      Profiles property to expose a single Profile object which has
      the expected property values.

Cq-Depend: chromium:2141443
Change-Id: I487fca97b026b3fce1c1096b969957659c62fe70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2142322
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>