dlcservice: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes this code
clearer and easier to modify, since if a case is added to this `switch`
later, the user should get a -Werror'ed warning if they forget to add a
`return`.

BUG=chromium:1066998
TEST=CQ

Change-Id: I02ba2c78ca11967e87fe00ed6ad695f37c7dd671
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2141241
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>