shill: cellular: change APN list behavior to merge APN details
Currently if the modem has a profile or provisioned context with
the same details as one in the mobile operator database, then that
APN will show up twice in the list and in the UI, possibly with
different names. Instead, collapse APNs for which connection
details are the same into one entry.

BUG=None
TEST=unit tests

Change-Id: I82f805cb9e83853d4b2c4cdf2d57a5e7774e827e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2142764
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>