crash: Ignore more errors due to lock_core_pattern
If lock_core_pattern is present and we fail to write to core_pattern,
crash_reporter --init fails, unless we're trying to disable
crash_reporter.
Extend this logic so that it also allows failures writing to
core_pattern if the core_pattern is already in the desired state.

Many crash reporter integration tests attempt to re-initialize crash
reporter to ensure it's in a valid state, and were spuriously failing
because core_pattern was locked.

Testing note: I don't have a device with a kernel <= 3.18 easily
available, so I haven't tested this change on anything that actually
uses lock_core_pattern.

BUG=chromium:1068641
TEST=cros_workon_make; tast run $DUT crash.KernelCrash.mock_consent

Change-Id: Icf9a10a3542ccb92bf6518a44652722aea37ee6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2142751
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>