diagnostics: add ProbeError to TimezoneInfo telemetry probe
This change modifies the TimezoneInfo probe to be a mojo union between
the TimezoneInfo and a ProbeError. This insures the result of the probe
can only be the data or an error. For each error condition that was
previously ignored, add a ProbeError.

In addition, factor out function to create a ProbeError and log the
message to a common error_utils file.

BUG=chromium:1041153
TEST=FEATURES=test emerge-sarien diagnostics (pass);
(dut) telem --category=timezone (shows timezone info)

Change-Id: I8a9191dc88c2f449f9c34711f062669ada0ebab4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2145475
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>
Tested-by: Trent Begin <tbegin@chromium.org>