camera: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes it clearer
that the above switch always returns. If a non-returning `case` is added
down the line, the dev should get a -Werror'ed warning about lacking a
`return` with a valid value.

BUG=chromium:1066998
TEST=CQ

Change-Id: I3716ef2f9d52a8f13516b60d7f705e475d3fc479
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2143271
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>