st_flash: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes it clearer
that the above code is an infinite loop. If someone does add a `break`
down the line, they'll get a -Werror'ed warning about lacking a
`return;` with a valid value.

BUG=chromium:1066998
TEST=CQ

Change-Id: Ib7af8cd26d49c7ef5d7bdf9dbf7b7792c081c5cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2143292
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>