vm_tools: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes it clearer
that the above code is an infinite loop. If someone does add a `break`
down the line, `main` always gets an implicit `return 0;` at the bottom
anyway.

BUG=chromium:1066998
TEST=CQ

Change-Id: If6664872f45e978ed94c4a137b01f253bc79331a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2143293
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>