diagnostics: add ProbeError to FanInfo telemetry probe
This change modifies the FanInfo probe to be a mojo union between a list
of FanInfo and a ProbeError. This insures the result of the probe can
only be the data or an error. For each error condition that was
previously ignored, add a ProbeError.

BUG=chromium:1041153
TEST=FEATURES=test emerge-sarien diagnostics (pass);
(dut) telem --category=fan (shows fan info)
(wilco dut) telem --category=fan (shows nothing)

Change-Id: I91472b8cbee76eab322c72e70e74edbab3d5804b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2145088
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>
Tested-by: Trent Begin <tbegin@chromium.org>