arc: networkd: Release fd from arc_networkd::Socket
The System-proxy project has an implementation of arc_networkd::Socket
that is a wrapper over a socket created by libcurl.

When the connection is closed, libcurl expects to close the socket
itself but fails with an error if the socket is closed due to the
underlying ScopedFD running out of scope.

BUG=chromium:1042626
TEST=none

Change-Id: I483befbbd0000eb3fa24ba5cc0a9a08a3ded70f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2144202
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>