dlcservice: Get user directories
Allow globally to get the device's user directory. This directory has
the list of user directories managed by Chrome. The name of each
directory is the sanitized username that we will be using for
ref-counting to get the current list of users on the device.

BUG=chromium:1069162
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: I07113e8d1355fcbb973263316529c88fb1375bb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2148246
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>