diagnostics: add ProbeError to MemoryInfo telemetry probe
This change modifies the MemoryInfo probe to be a mojo union between a
list of MemoryInfo and a ProbeError. This insures the result of the
probe can only be the data or an error. For each error condition that
was previously ignored, add a ProbeError.

BUG=chromium:1041153
TEST=FEATURES=test emerge-sarien diagnostics (pass);
(dut) telem --category=memory (shows memory info)

Change-Id: I8ddfed8b379538feefa48703b5402d7e6fcb6f88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2148096
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>
Tested-by: Trent Begin <tbegin@chromium.org>