diagnostics: add ProbeError to BacklightInfo telemetry probe
This change modifies the BacklightInfo probe to be a mojo union between
the BacklightInfo and a ProbeError. This ensures the result of the probe
can only be the data or an error. For each error condition that was
previously ignored, add a ProbeError.

BUG=chromium:1041153
TEST=FEATURES=test emerge-sarien diagnostics (pass);
(dut) telem --category=backlight (shows backlight info)

Change-Id: I4bcfa8f7dff87886c5473aa8d59bb66371501aec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2148241
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Tested-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>