shill: Merge ModemManager into ModemInfo
ModemInfo is just a light wrapper around ModemManager,
which doesn't really managed anything and is something
of a confising name.

This CL merges the two classes, using the ModemInfo
name since it is already used in a number of other places,
and is more consistend with DeviceInfo and other similar
classes.

This includes a bit of minor cleanup as well.

BUG=b:153675341
TEST=cros_workon_make --board=${BOARD} --noreconf --test shill

Change-Id: Iab69a66e85b894157a7506f7a2719e5dd7b89f3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2148223
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>