dlcservice: Add and Handle DlcState::MOUNTABLE
The |DlcState::MOUNTABLE| is a state a DLC is in when it is marked
mountable, but not mounted yet. This allows during the startup of
dlcservice to determine which DLCs are mountable.

Dlcservice receives a DBus method call from update_engine whenever there
is an install/update completion. Dlcservice will leverage the persisted
verified prefs prior to mounting the DLC images and transitioning to a
|DlcState::INSTALLED|.

BUG=chromium:1059126
TEST=FEATURES=test emerge-$B dlcservice dlcservice-client

Cq-Depend: chromium:2146104
Change-Id: Ice41c5e1d155fe72f50e7ba330e85a28dff944c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2148224
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>