hermes: apdu_test: Avoid EXPECT_FRAGMENT UB
Apdu::GetNextFragment will not set the output parameter when there is
no remaining fragment. Because of this, tests were relying on
UB when EXPECT_FRAGMENT was called with no additional arguments (used
to indicate that there was no additional fragment). Instead of doing
that, we simply explicitly check for HasRemainingFragment.

BUG=chromium:847619
TEST=All unit tests are passing.

Change-Id: I4c8fc8048b969dc04b21367d3aaa51bac1f386bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2149337
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>