diagnostics: add ProbeError to CachedVpdInfo telemetry probe
This change modifies the CachedVpdInfo probe to be a mojo union between
a list of CachedVpdInfo and a ProbeError. This ensures the result of the
probe can only be the data or an error. For each error condition that
was previously ignored, add a ProbeError.

BUG=chromium:1041153
TEST=FEATURES=test emerge-sarien diagnostics (pass);
(supported dut) telem --category=cached_vpd (shows vpd info)
(unsupported dut) telem --category=cached_vpd (shows NA)

Change-Id: I3a12f445992641cea374b037b9162412c02cac9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2148820
Tested-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>