shill: l2tp: Document public D-Bus properties
The L2TPIPsec.* D-Bus properties have never been documented. This
change adds documentation for all of the public properties (there exist
a few properties whose names are defined in an anonymous namespace in
l2tp_ipsec_driver.cc which seem to exist primarily for development
purposes and should not be set by shill clients in their current
state).

Note that, as for all VPN properties, the properties are never
readable directly. Those that are not write-only may be read by
getting the kProviderProperty, which maps between property name and
value. Hopefully this hack can be eliminated one day, but for now,
even properties that are semantically readwrite will be documented as
writeonly (consistent with how OpenVPN properties are documented).

BUG=None
TEST=None.

Change-Id: I632b1512172844f92a77dde521ed0ca1b536ee96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2149353
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>