cryptohome: Move timestamp handling for mount into separate file
Move the timestamp handling from vault keyset in
Mount::UpdateCurrentUserActivityTimestamp into a separate file.
This will make it easier to remove the keyset file securely when
the user is removed or the machine is powerwashed.

When loading the vault keyset, load the timestamp from the timestamp
file if it exists.

BUG=chromium:1056465
TEST=FEATURES=test emerge-hatch cryptohome

Change-Id: Ie5188cf7db7eb0f3ec8a5168809dc0a30e9cd762
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2149867
Tested-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>