dlcservice: Add user refcounting
Currently, when a DLC is uninstalled, we immediately remove it. But what
if another user of the device was still using this DLC? It wouldn't be
fair to them to remove something that they were using. So we need some
sort of reference counting and garbage collection for DLCs that have
dangling references. This CL adds two types of such reference counting
objects. One for system DLCs (DLCs that have no specific device users)
and one for DLCs which are installed per request of the device users. Both of
these kinds will have an expiration delay of 5 days. If no one was using
these DLCs after the 5 days delay, dlcservice will remove the DLC files
and free up space.

Also we trigger the garbage collection after the first half hour of the
start up and once everyday. The reason is if the user hasn't opened
their device for a long time (more than 5 days), we will give them a
chance to install it if they need to.

BUG=chromium:1069162
TEST=cros_workon_make --board reef --test dlcservice
TEST=dlcservice_client --install dummy-dlc
TEST=tast run chromeos6-row4-rack9-host15.cros platform.DLCService

Change-Id: Ibe7534eb35dc7a4b907999c765a49f40e69c5e35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2157665
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>