shill: Cleanup CelluarService methods
The CellularService methods are a bit scattered and the header
does not match the .cc file. This CL fixes the order and groups
overrides.

This is in preparation for some significant changes to make
CellularService behave more like other Service classes, including
the addition of a provider class.

BUG=b:154014577
TEST=cros_workon_make --board=${BOARD} --noreconf --test shill

Change-Id: I89cabec0937e3fa6d1854bd067f72458fd027641
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2151014
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>