shill: Use IMSI to identify CellularService entries
Currently we use a variety of possible properties to generate
identefiers for CellularService entries in the profile. These
are different than the actual properties we use to match entries
(either IMSI or MEID).

It is desired that CellularService entries should be identified
exclusively by IMSI. (If no IMSI is available we should not be
able to connect, so we should not have a Profile entry anyway).
This changes the identifier to always be IMSI.

This also adds a test to ensure that existing profile entries with
a matching IMSI will still be loaded even though the identifier
is different (existing behavior). It also cleans up the tests.

BUG=b:154014577
TEST=emerge-${BOARD} shill && cros deploy localhost:2222 shill

Change-Id: I601a71116af2ffd9190a55ef9cfdf4ff287ca698
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2151020
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>