dlcservice: Fallthrough create
When we try to install a DLC that already exist, we need to make sure
all files/directories associated with that DLC is already created if
not exist. Allowing creation to be fallthrough also let us not to delete
a DLC when the verification failed. We just override it. This allows us
not to send a signal for deletion when the DLC is going to be installed.

This also allows us to include directories/files that we want for DLC to
be added without worrying about backward compatibility and creating
speciall cases (like the current pref directories).

BUG=chromium:1071261
TEST=cros_workon_make --board reef --test
TEST=dlcservice_utils install --id dummy-dlc && dlcservice_utils install --id dummy-dlc

Change-Id: I346e101bc1b2fefb506880163214010416f068e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2208372
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>