shill: Fix mock warnings in CellularServiceTest
This is in preparation for some CellularService changes.

Currently the tests spam a lot of interesting mock warnings,
making the test results difficult to parse and errors difficult
to notice.

This cleans up CellularServiceTest to add a common Metrics
expectation to the mock and a common Manager expectation to
CellularServiceTest.

BUG=b:154014577
TEST=P2_TEST_FILTER="CellularServiceTest.*" cros_workon_make \
     --board=${BOARD} --noreconf --test shill
     (Should have no mock warnings)

Change-Id: Ic3a10884ce8f5db6019a4daa373d09290e9b8818
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2152970
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>