camera: Start cros-camera service when USB camera connected
In some cases, the camera might be probed late by kernel, such as ~20s
after boot as shown in http://b/151287253#comment39.

This CL adds a new udev rule to emit an upstart event
usbcam-device-added when an USB camera is added, and starts the camera
service accordingly if it's not running nor respawning.

Note that we still rely on the respawn mechanism for the normal case
where the camera(s) are probed early, because we need to wait until all
built-in cameras probed before starting the service to assign the camera
id properly.

BUG=b:154262919,b:151287253
TEST=Simulate the situation by an external camera.

Change-Id: Iaa8e1e4063b07ae1c273e48639a4749c899e6bc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2154290
Tested-by: Shik Chen <shik@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>