debugd: accept language when starting Plugin VM dispatcher
We need to pass user's selected UI language to the Plugin VM dispatcher
so that product can localize its UI to match the rest of the system. To
achieve this we will be setting CROS_USER_UI_LANG environment variable
in the upstart unit that starts the dispatcher.

Note that we do not support locales in crosh sessions (LC_ALL is set
unconditionally to en_US.utf8) so we hardcoding "en-US" when starting
the dispatcher from "vmc".

BUG=b:150910107
TEST=Switch Chromebook to Russian, start VM, examine dispatcher's
environment.

Change-Id: Iefebd9c2a5d21ae20042b1360677ba231ceba444
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2155189
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>