shill: Eliminate unused device policy code
This code was added to load throttling policies on startup:
https://chromium-review.googlesource.com/c/aosp/platform/system/connectivity/shill/+/386364

However it was never implemented. Chrome sets the throttling.
Presumably this was added with the intention of ensuring that
throttling was set before Chrome started, however in practice
that should be a very small window with no significant traffic.
We also currently rely on Chrome for loading and verifying
policies, so until/unless we change that, we are better off
relying on Chrome to set these values.

BUG=chromium:1070733,chromium:634529
TEST=unit tests

Change-Id: I8cbcd2c48c8e755f92cc3b6e4b37568ef79c4054
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2154296
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>