chromeos-config: make unspecified smbios-name-match match any SMBIOS name
Platforms like Betty (a VM) have arbitrary and useless SMBIOS names,
depending on the underlying emulated firmware.  On these platforms,
it's useful to specify no smbios-name-match and match any SMBIOS name.

This is the libcros_config side of the change to make this work.  A
few more changes to cros_config_schema will be required to allow
this.

BUG=b:154129827
TEST=unit tests

Change-Id: I0d4a320377bc1c598cd3322bb606b3ca69512592
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2159765
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>