camera: Ensure the environment is ready when starting cros-camera service
If the usbcam-device-added event is triggered too early, the environment
such as the /run/camera directory might not be ready yet. Ignore the
usbcam event in that case and wait "starting system-services" event to
start cros-camera again when the environment is ready.

The logic is mostly adapted from hammerd [1].

Note that "started" in upstart is an event, not a state [2]. So
`started system-services and usbcam-device-added` will not work as
expected in the "start on" stanza.

[1] http://cs/chromeos_public/src/platform2/hammerd/init/hammerd.conf?l=48&rcl=4d215c8e69c9d524f0b7dca0146c935ab5efdf05
[2] http://upstart.ubuntu.com/cookbook/#events-not-states

BUG=b:154668367,b:154262919,b:151287253
TEST=Simulate the situation manually.

Change-Id: I611c7346a5c64e2b3d2e139061838d3041431c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2160292
Tested-by: Shik Chen <shik@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>